fix(jobs): local runner for enterprise (#2212)
## Describe your changes
Issues with the runner failing health checks for enterprise. Using a a
remote runner requires the API call from jobs to the runner stay open
for the duration of the sync. This can be problematic as the connection
can drop or be terminated for a number of reasons. This piece of
[code](https://github.com/NangoHQ/nango/blob/master/packages/jobs/lib/integration.service.ts#L124-L133]
should be refactored soon so that this logic is more portable across
different cloud providers besides just our own.

## Issue ticket number and link

## Checklist before requesting a review (skip if just adding/editing
APIs & templates)
- [ ] I added tests, otherwise the reason is: 
- [ ] I added observability, otherwise the reason is:
- [ ] I added analytics, otherwise the reason is: