feat(logs): pagination and infinite scroll (#2213)
## Context

Fixes NAN-1004
Fixes NAN-997
Contributes to NAN-908

Took me a day just to handle this :|
It was much more painful than anticipated for many reasons: 
- plenty of filters to combine
- live refresh that prepend and also updates the current rows
- infinite scroll that append but only if nothing else is changing in
the UI

When filtering the cursor changes, but when something change too fast:
race condition can happen, so I had to use ref and tricks. I also had to
aggregate rows into an array from which I could: update, append and
prepend; I had to handle writing in this array from multiples places
which in React is not easy.
The result is honestly quite flaky and very hard to grasp, I don't give
it 2 weeks until something is broken.
Ideally we should remove at least the "update", only query before or
after the cursor, and update the states of the rows with websockets.

## Changes 

- **Add cursor** to searchOperations

- **Handle infinite scroll and live refresh for operations**

- Fix layout scroll box
The DashboardLayout component was a mix of fixed and flex box which made
the scroll area not consistent. For example a page without content would
still have a scrollbar. I also added an option to allow a content to
stretch the full width.

## Test

**(I added some fake lag to help me spot loading)**


https://github.com/NangoHQ/nango/assets/1637651/3b4256b6-31ea-47e1-9004-07ee85b9c384