fix(logs): first prod feedback (#2224)
## Changes

Directly related to implem:
- UI: bad loading state when opening the operation drawer
- UI: search in operation's logs was not working after latest PR 
- ES: if we don't `refresh=true` index after inserting operation it can
sometimes be missing from the index when we reach jobs making it
impossible to update state, unfortunately that will slow down insert of
operation not sure by how much

Other related changes:
- `syncRunService` clarify what we expect in terms of logging options
(with types) when dryRun or not
We either `writeToDB` and needs everything related to logs, or nothing. 

- `syncRunService` now takes a mandatory logCtx (if writeToDb=true)
In `activities.ts` we create an operation for `sync run`, but when we
entered this service (on the same machine) we had to get the operation
back with the activityLogId which was definitely unnecessary and slowing
down everything.

- `triggerAction()` never takes an optional activityLogId and
`writeLogs` at all
It helps simplify the code and ensure we have the data in activities.ts.
I could not find a place where we don't pass an activityLogId and use
this writeLogs=false. I saw it was introduced in an old
[PR](https://github.com/NangoHQ/nango/pull/1202) let me know @khaliqgant