feat(pkgs): create kvstore  (#2235)
## Describe your changes

Fixes NAN-1071

- Create a reusable package for KVStore
Duplicated for now, can't remove it entirely from `shared` because
FeatureFlags and Locking, but it's a first step. I stopped exporting it
so it's no longer used by external package and I'll do a second pass to
remove it entirely.
Note: I did that because logs can't depend on shared


- Use it in Jobs


- Use it in Logs
As discussed orally, the main id is to be able to retrieve the indexName
quickly without doing a full search on the cluster. Maybe it's the fix
maybe not.