fix(database): [nan-1063] migrate to database package (#2236)
## Describe your changes
As part of another ticket related to webhooks, I want to make a specific
webhook package. I don't want the webhook package to depend on shared
but shared contained the database as of now. This PR moves the database
connection logic to its own package freeing it from shared and making it
easier to use freely in other packages now. Note that the seeders have
to stay due to the usage of services.

## Issue ticket number and link
Contributes to NAN-1063

## Checklist before requesting a review (skip if just adding/editing
APIs & templates)
- [ ] I added tests, otherwise the reason is: 
- [ ] I added observability, otherwise the reason is:
- [ ] I added analytics, otherwise the reason is: