fix(orchestrator): race condition in dequeue
another race condition in orchestrator dequeue where the new task event
can be received at the same time of the timeout triggering. The dequeued
tasks will then not be sent back to the processor.
Here again the fix is simple: only try to dequeue after having cleanup
the timeout and check that no response was already sent