fix(orchestrator): race condition in dequeue (#2261)
another race condition in orchestrator dequeue where the new task event
can be received at the same time of the timeout triggering. The dequeued
tasks will then not be sent back to the processor. Here again the fix is
simple: only try to dequeue after having cleanup the timeout and check
that no response was already sent

## Checklist before requesting a review (skip if just adding/editing
APIs & templates)
- [ ] I added tests, otherwise the reason is: I managed to write a test
to trigger the race condition but it was only occuring 50% of the time
:(
- [ ] I added observability, otherwise the reason is:
- [ ] I added analytics, otherwise the reason is: