fix: missing created task events
when tasks are being scheduled by the scheduling worker a message is
posted to the scheduler that search for the tasks and emit a CREATED
event per task.
The message was posted within a transaction which led to scneario where
the scheduler search query would returns nothing because it would be
executed before the transaction was committed and the tasks actually
visible.
This commit modifies the scheduling worker so the message is posted
once the transaction is finished/committed