fix(logs): insert message without an ID (#2360)
## Describe your changes

Fixes
https://linear.app/nango/issue/NAN-1244/insert-messages-without-an-id

Now that we have Elasticsearch in prod I can focus a bit more time on
performance quick wins.

- Insert messages (not operation) without an ID
I knew this from the beginning but I just wanted to simplify the code.
They recommend not using an ID because if you pass one, all the
instances need to search in their whole catalog to see if this ID
already exists. Whereas if you don't send it, they will autogenerate an
ordered id without checking. This should help but no clue how much.

- Clarify types between OperationRow and MessageRow which were different
already but a bit mixed up
It's not perfect yet since most fields are optional 
- Fix some type errors