fix(server): log request and add route to tracing (#2371)
## Describe your changes

Fixes
https://linear.app/nango/issue/NAN-1250/set-httproute-in-express-tag

- Log server requests
It's a bit annoying to not see if the request has reached the backend,
especially in docker mode. Follow what Thomas has done for persist and
orch + colors and log level

- Add correct to tracing
Datadog doesn't instrument correctly (probably laziness since OT is
doing it right). The `route` prop is only available after match and the
tracer is also only avalaible at this stage. So there is only one place
to set this tag (`router.handle`), so I'm using asyncWrapper to not have
to instrument anything. The drawbacks is that only new endpoints using
this function are correctly sending the route.

This finally allow seeing the route directly in traces and being able to
graph/group routes by pattern

--- 
*Notice the holes are the old endpoints not using asyncWrapper*
![Screenshot 2024-06-19 at 16 19
57](https://github.com/NangoHQ/nango/assets/1637651/61244696-2fd9-4068-ab57-5b81de92e0a9)