chore: add last_scheduled_task_id column to schedules
DueSchedule query is too slow because for each schedules we are going
through the tasks table to find out if there is any running tasks or if
any task has been running recently.
In order to make the query faster we are adding a last_scheduled_task_id
column to schedules which is populated when scheduling a task. This
column will allow us to join schedules with tasks based on this column
and look at the state and created_at of the tasks to find due schedules