fix: update schedule last task when sync is triggered manually
without it we can run into scenario where a sync is triggerd manually
right before the sync due date. Orchestrator scheduling will then not
know that a sync is already running for this schedule and will trigger a
new sync to start, leading to 2 syncs running at the same time.