fix(metrics): add sync, action, webhook, post script, proxy (#2398)
## Describe your changes

Fixes
https://linear.app/nango/issue/NAN-1279/track-new-metric-and-log-details-in-datadog

We only have logs to compute those metrics, and we want to have more
than 2 weeks history. It's a bit messy because the obvious place is to
put that in `processor.ts` or `hander.ts` but then it would not failure
that happen before exec (i.e: at orchestration). But then sync does not
get orchestrated the same way, so a bit annoying. And then early return
makes stuff a bit more repetitive.

- Add metric for sync execution
- Add metric for action execution
- Add metric for post script connection execution
- Add metric for incoming webhook
- Add metric for proxy
It's the more messy since we have dozen of early return conditions
- Fix typo in Reconcile