fix(sdk): handle internal issue when reaching persist (#2416)
## Describe your changes

Fixes
https://linear.app/nango/issue/NAN-1304/persist-calls-sometimes-fails

- Handle ECONNREFUSED and ECONNRESET when reaching Persist
Right now a call that failed like this would break the script with an
ugly error, it just makes it more readable and easily identifiable by
us. I have noted that a better solution would be to retry on top of
that. This can happen when we deploy or when the service scale down.
Also I have decided to make logs less critical by failing "silently"
unfortunately there is no way right now to warn without stopping the
script and we can't log because it just failed :D Let me know if you
think it's okay