fix(nango-yaml): handle deep circular ref, correctly output recursive models (#2448)
## Describe your changes

Fixes
https://linear.app/nango/issue/NAN-1331/handle-deep-recursive-model-and-output

- **Handle deeply nested circular ref**
While testing, I encountered an edge case I initially spotted but not
tested hard enough. I added supported for circular ref, but only if it
was a direct circular ref (`A -> A`); but well it's possible to go
deeper, i.e: `A -> B -> C -> A`. It's now supported.

- **Output all used models per scripts** correctly
With deep circular ref and other large models, I wasn't fetching the
model names recursively correctly.
It wasn't the case before the refacto but the `model_schema` was only
used for display; Now I'm using that field as a source of truth to find
related json schema, so in some case we were missing a definition.