feat(integration-templates): [nan-1328] add xero templates (#2452)
## Describe your changes
Add in xero templates. Please note that I didn't write this code so
don't have a lot of context on it. I did however, add a post connection
script to fetch the tenant id and move the logic to fetch that into a
separate call since it was repeated. Also abstracted out some logic for
mappers where appropriate.

## Issue ticket number and link
NAN-1328

## Checklist before requesting a review (skip if just adding/editing
APIs & templates)
- [ ] I added tests, otherwise the reason is: 
- [ ] I added observability, otherwise the reason is:
- [ ] I added analytics, otherwise the reason is: