fix: hard delete one schedule at a time
Deleting a schedule would cause the cascading deletion of all its tasks
In order to avoid massive deletion of tasks that could be slow we limit
the deletion of schedule to one at a time.
Schedules are currently marked as deleted very infrequently so this
deletion rate is still more than enough.