fix: token refresh locking logic
One customer is experiencing connection refresh failures with the
following error:
Acquiring lock for key: lock:refresh:XXX timed out after 12000ms
They are making concurrent calls to `/proxy` while the token is expired.
Each request trying to acquire the lock is waiting and then attempts
again to refresh the token. When a lot of concurrent requests, waiting
for all the other requests in the queue to refetch the connection token
can exceed the lock acquisition timeout.
This commit is introducing 2 fixes:
- after waiting for acquiring the lock, first check if the connection
  has been refreshed, if yes returns the credentials directly.
- if lock acquisition times out, also check if the connection has been
  refreshed and only fails if it is not the case.