fix: token refresh locking logic (#2494)
One customer is experiencing connection refresh failures with the
following error:
Acquiring lock for key: lock:refresh:XXX timed out after 12000ms They
are making concurrent calls to `/proxy` while the token is expired. Each
request trying to acquire the lock is waiting and then attempts again to
refresh the token. When a lot of concurrent requests, waiting for all
the other requests in the queue to refetch the connection token can
exceed the lock acquisition timeout.
This commit is introducing 2 fixes:
- after waiting for acquiring the lock, first check if the connection
has been refreshed, if yes returns the credentials directly.
- if lock acquisition times out, also check if the connection has been
refreshed and only fails if it is not the case.

## Issue ticket number and link

https://linear.app/nango/issue/NAN-1164/salesforcebox-connections-failing-to-refresh-intermittently

## Checklist before requesting a review (skip if just adding/editing
APIs & templates)
- [ ] I added tests, otherwise the reason is: 
- [ ] I added observability, otherwise the reason is:
- [ ] I added analytics, otherwise the reason is: