fix(validation): cache ajv compilation (#2502)
## Describe your changes

Fixes https://linear.app/nango/issue/NAN-1418/cache-ajv-data-validation

- Cache AJV compiled schema 
We saw an increase in batchSave latency, initially we thought there was
an issue with the request but it's the records validation that was to
blame. We compiled the schema at every record which was not a good idea.
In this PR we now cache schema per `model-version` which allows stable
cache and automatic invalidation on deploy. There is no automated cache
clear strategy as everything is happening in the runner/cli, but it's
definitely something to keep in mind if we use it somewhere else.
This drastically improve performance (or get it back to what it was
before):
for 10K large records `42s -> 4s`