fix: update sync frequency
Updating frequency from the integration UI is failing.
The endpoint retrieves all the syncs for the given integration
and attempt to update the schedule frequency in the orchestrator.
However the query to retrieve the syncs was wrongly returning sync that
didn't belong to the current user because the join between syncs and
sync_configs was done only on the name which is not unique.
This commit fixes the query by also joining sync and connections to only
select rows where connection's environment_id matches.
Note: We should eventually add a foreign key to join syncs and sync_configs table