feat: expose env.js from backend to frontend (#2540)
## Describe your changes

Fixes
https://linear.app/nango/issue/NAN-1487/expose-envjs-to-pass-env-variables-from-back-to-front

### Major changes

- Expose a fake `env.js` from backend to frontend
The goal is to be able to pass feature flags and env var dynamically
before react loads. The ultimate goal is to remove the need for specific
compilation to enable/disable some features and also adapt the UI based
on the flags while being in sync with the backend.
I haven't done a full pass yet but it showcase the possibilities. Once I
remove the posthog and sentry keys we can stop compiling
production/staging/enteprise docker images.

### Minor changes
- Fix cookie locally
I added flag to secure cookies but it's not possible to secure them in
`http` or `localhost`.

- Removed some unused images
- Enable static assets caching
Not sure why it was disabled, let me know.


### 🧪 How to tests?

> NB: It's currently in staging

- Locally: set `FLAG_AUTH_ENABLED=false|true`
`FLAG_MANAGED_AUTH_ENABLED=false|true` to enable/disable stuff in the UI