fix(cli): [nan-1145] only compile and deploy single sync (#2582)
## Describe your changes
Deploying by passing in a single sync or action shouldn't compile all
the other files and should only deploy the specified file. Right now it
compiles all files and if a sync is passed in it tries to deploy all
other actions. This PR fixes both issues

## Issue ticket number and link
NAN-1145

## Checklist before requesting a review (skip if just adding/editing
APIs & templates)
- [ ] I added tests, otherwise the reason is: 
- [ ] I added observability, otherwise the reason is:
- [ ] I added analytics, otherwise the reason is: