feat: Integration UI revamp (#2594)
## Describe your changes

Fixes
https://linear.app/nango/issue/NAN-1542/script-v2-integrate-settings
Fixes https://linear.app/nango/issue/NAN-1537/script-v2-integrate-ui
Fixes
https://linear.app/nango/issue/NAN-1429/editing-an-integration-name-to-an-existing-integration-silently-errors

### Major Changes

- Rework `PATCH /flows/:id/enable`
- Rework `PATCH /flows/:id/disable`
- Rework `PATCH /flows/:id/frequency`
- Rework `POST /integrations/`
- Rework `GET /integrations/:id`
- Rework `PATCH /integrations/:id`
- Rework `GET /integrations/:id/flows`

- Rework UI /integrations/:id 
New endpoints page, settings page revamp (and split components per
credentials type), new endpoint page. The goal of this revamp is to
present integrations as a list of endpoints rather than a list of
scripts.


### 🧪 How to review and test?

This PR is massive since it touches a lot of components and endpoints,
sorry about that. The best way to review that is to begin with the UI
and check the code. I have extensively tested it, but there is a lot to
go over and there are probably some edge cases.

- Create a new integration
- Enable / Disable actions and syncs
- Edit integration credentials (if you can test: try doing it for
multiple types of credentials)
- Check one endpoint, access the URL directly,
- Change frequency of an endpoint 
- Delete an integration

> [!NOTE]
> UI/UX has been reviewed with Bastien but we are not sure about
everything, feel free to raise confusing points or things that could be
improved

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Introduced new endpoints for managing integrations, including
retrieval of integration flows and details based on configuration keys.
	- Added a customizable Select component for enhanced user interactions.
- Enhanced the user interface with new components for displaying and
managing API endpoints.
- Implemented a new endpoint for updating synchronization flow
frequencies.

- **Improvements**
- Enhanced validation logic for various parameters across multiple
components.
- Updated API specifications to reflect new response structures and
improved user feedback mechanisms during integration operations.

- **Bug Fixes**
	- Corrected minor issues in error messages and validation checks.

- **Chores**
	- Removed outdated API utility functions to simplify the codebase.
	- Added new dependencies to support enhanced UI components.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->