fix(syncs table): sync_config_id must be updated when new deploy (#2736)
I introduced a bug last week when adding the `sync_config_id` column to
the `syncs` table. The value is set when a sync is created and never
changes. However each time a sync definition is deployed (or prebuild
templates are upgraded) a new version of the `sync_config` is added but
the `sync.sync_config_id` is currently not updated and keep pointing to
the same `sync_config` that is now inactive.

Impact: This new `sync_config_id` is only used by the patch frequency
endpoint and demo auto idle feature for now but the bug should be fixed
asap

This commit is ensuring that sync_config_id in syncs table is updated on
deploy/upgrade as well as fixing the drift of `sync_config_id` values in
the db