fix(api): GET /integrations + GET /integrations/:uniqueKey (#2744)
## Describe your changes

Fixes https://linear.app/nango/issue/NAN-1746/get-integrations

- Add `GET /integrations`
- Add `GET /integrations/:uniqueKey`
I changed the param name to `uniqueKey` it's what we return in the
payload, happy to discuss.
I didn't add `?include_creds` like in the previous endpoint, when I
announce the deprecation, it will be noted as a difference, and I want
people to reach out if necessary. This is leaking stuff, and it would be
great if we could get rid of it. I actively reached out to the only
person I could identify in the list, but no customer is currently using
this feature.

- Deprecate `GET /config` and `GET /config/:providerConfigKey`
Note that there are more endpoints under than prefix.


> [!NOTE]
> There is a little bit of code duplication but that's expected so we
don't break legacy

---

Open questions: I haven't added/modified the node-client, I have 2
solutions:
- rename `getIntegration` to `getIntegrationLegacy` but that would force
a migration as soon as they upgrade (or use new name)
- add a new method but I don't know what name should I use.