feat: make connectionId optional when creating a new connection (#2746)
## Describe your changes

Making connectionId optional when creating a connection. If connectionId
is not provided a uuid is generated.
I haven't updated the documentation which is still relevant. I will do
it once we decide not passing the connectionId is the default way. With
this PR we are making it possible without promoting it yet.

## Issue ticket number and link


https://linear.app/nango/issue/NAN-1747/make-connectionid-optional-when-creating-a-new-connection

## Checklist before requesting a review (skip if just adding/editing
APIs & templates)
- [ ] I added tests, otherwise the reason is: 
- [ ] I added observability, otherwise the reason is:
- [ ] I added analytics, otherwise the reason is: