feat(sdk): add new endpoints (#2747)
## Describe your changes

Fixes
https://linear.app/nango/issue/NAN-1753/document-new-endpoints-integrations-and-providers

- Add new endpoints to the SDK
There is only one method that was different enough, queries and output
that even with signature overload it wasn't possible to make it
non-breaking. Honestly, I didn't find anything compelling to deprecate
without either being ugly or breaking customers' scripts, so very open
to suggestions.

- Document new SDK method
- Add support for `include=webhook` in `GET /integrations/:key`
The SDK needs it, and some customers use it. It's a secret but not
credentials, so I didn't put a too vague name. If we need credentials at
some point, we can differentiate.