feat: add keystore package (#2756)
## Describe your changes

A new keystore package containing logic to create, store, retrieve and
delete private keys.
I decided to make it an independant package instead of adding it to
`server` so it can be self-contained and reusable. It means the table
migrations are different (which creates new migrations table) but also
means the package can be used independently.
The db client must be provided though so it can use the same as the one
used by the service importing `keystore`.
In this PR I added keystore to `server`. In next PR I will add the
endpoint used in the new auth flow to create and check session token

## Issue ticket number and link
https://linear.app/nango/issue/NAN-1758/implement-keystore-package

## Checklist before requesting a review (skip if just adding/editing
APIs & templates)
- [x] I added tests, otherwise the reason is: 
- [x] I added observability, otherwise the reason is:
- [ ] I added analytics, otherwise the reason is: