fix(jobs): try to handle knex pool exhaustion (#2771)
## Describe your changes

- Handle unhandledRejection and uncaughtException
This is something we need to do globally, but it's the opportunity to
start. Not entirely sure the error we have is related to uncaught error
but we'll see. We don't properly log errors anywhere because we don't
have a working sentry and our Datadog error log is shaky.

- Add proper health checks
Turns out /health is only called on web service, so our current health
check (which was not doing much) was useless anyway. I try to query
nothing and timeout quickly

- Remove transaction from Delete Syncs cron
This is the only place where there was an exotic way of using
transactions. Honestly, I'm not convinced there was an issue but since
the transaction was useless I'm removing it