feat(connect): require session token to load (#2787)
## Describe your changes

Fixes https://linear.app/nango/issue/NAN-1822/enforce-sessiontoken-usage

- Enforce session token to load UI
We request the sessionToken, open the modal, the modal sends "ready"
when react is loaded, and the parent sends back the sessionToken. It's
not passed in the UI to avoid leaking the token in logs.

- Add `POST /api/v1/connect/sessions`
We needed a quick way to generate this token for the dashboard, it"s
better to assume the UI do not have access to the private key (it's the
case but that's not good imo). This endpoint is actually calling the
public one by faking the request.