fix: 'window is already open' bug
Closing the auth popup and calling nango.auth again result in the
`window is already open` error. It doesn't prevent the new popup to open
but it is visually annoying and confusing.
With this commit, before opening the popup again we check if previous
one is closed and clear the state