feat(connect): open beta (#2809)
## Describe your changes

Fixes https://linear.app/nango/issue/NAN-1853/productize

- Enable UI in cloud env
- Allow UI to open Connect UI in an iframe
- Less hardcoding of api hostname
Still not good though, it will be impossible to not pass an env var at
compile time. It's the same issue for the current dashboard, the trick I
used to load dynamic env with `/env.js` is only working because it's
loaded in the same subdomain. I guess that means we'll still need to
build two docker image staging + prod unless we remove frontend from the
docker image.

- Clear open popup on early exit


It's currently accessible in staging