fix: out of sync between task state and sync_job status
When user currently cancels a task we only update the sync_job entry if
the abort task is successfully processed by job.
However if job for some reason is having issue, the customer is left
with seeing the task still running in the dashboard.
With this commit the sync job is always updated when cancellation is
clicked and doesn't depend on Job actually aborting the task. Even if
jobs the scheduler will eventually recover by cancelliing/timing out the task