fix: out of sync between task state and sync_job status (#2829)
When user currently cancels a task we only update the sync_job entry if
the abort task is successfully processed by jobs.
However if jobs for some reason is having issue, the customer is left
with seeing the task still running in the dashboard. With this commit
the sync job is always updated when cancellation is clicked and doesn't
depend on Job actually aborting the task. Even if jobs is misbehaving
the scheduler will eventually recover by cancelling/timing out the task

## Issue ticket number and link

https://linear.app/nango/issue/NAN-1865/fix-sync-status-out-of-sync-on-connections-page

## Checklist before requesting a review (skip if just adding/editing
APIs & templates)
- [ ] I added tests, otherwise the reason is: 
- [ ] I added observability, otherwise the reason is:
- [ ] I added analytics, otherwise the reason is: