fix(connections): update last fetched at to be 33% (#2838)
## Describe your changes
Last fetched at is used to to check for token refresh. Reduce the
setting of this value to 33% of the time. This will mean that more
connections will be refreshed in the cron but since this call is made
every time a connection is retrieved even if it isn't refreshed is a
little excessive. Every time the credentials are actually refreshed then
update last fetched at.

## Issue ticket number and link

## Checklist before requesting a review (skip if just adding/editing
APIs & templates)
- [ ] I added tests, otherwise the reason is: 
- [ ] I added observability, otherwise the reason is:
- [ ] I added analytics, otherwise the reason is: