chore: hard delete sync jobs when a sync is deleted (#2834)
Sync jobs table is growing rapidely and could become a problem at some
point. The only advantage of doing a soft delete is debugging but we
rarely need it. We could also have a more complex 2 steps delete
(immediate soft + hard later) but I don't think it is worth it.

## Issue ticket number and link

## Checklist before requesting a review (skip if just adding/editing
APIs & templates)
- [ ] I added tests, otherwise the reason is: 
- [ ] I added observability, otherwise the reason is:
- [ ] I added analytics, otherwise the reason is: