fix: lock can be released even when not holding the lock (#2858)
## Describe your changes

I wanted to use the redis based locking mechanism to refactor the
refresh tokens logic but I found out a pretty big flow where a lock can
be released by a process that doesn't hold it. That's a pretty bad
design flow but I am not sure of magnitude of the impact since we are
using it only to prevent concurrent token refresh and the issue would
only materialize in the very specific scenario where the lock holding
process is still trying to refresh the token after the lock timeout
(10s)

Anyway this fix ensures that to release it, a lock must first be
acquired.
I have also moved the instantiation of the kvstore so it can be reused
outside of the connection service. (I am gonna use it in the
refreshTokens cron job)

## Checklist before requesting a review (skip if just adding/editing
APIs & templates)
- [ ] I added tests, otherwise the reason is: 
- [ ] I added observability, otherwise the reason is:
- [ ] I added analytics, otherwise the reason is: