fix(connection): getConnectionCredentials should return credentials even on error (#2870)
## Describe your changes

Fixes
https://linear.app/nango/issue/NAN-1933/fix-ui-crash-when-loading-a-connection-with-token-refresh-error

- Split getConnectionCredentials and initial connection fetch
For the /connections/:id endpoint we still want to return the connection
even if the refresh failed but if there was an error we stripped the
credentials to avoid leaking stuff. Initially, I had just reverted this
and stripped the credentials outside of the function at multiple places
but the code was even more crappier.
So I split the function, this allows us to not repeat call to fetch the
integration on the API and correctly output the errors. It's a bit less
straightforward but more granular