feat: add OpenTelemetry export
Log operations are sent as OpenTelemetry traces to a otlp backend set by
customers on the environment level
It is using a custom SpanProcessor from opentelemetry-js lib
with one exporter per customers. I assume this feature will only be
available to scale plan customer and therefore the number of exporters
will stay low
Another thing to note is that operation must be started and ended on the
same service, which is the case today but could become a limitation in
the future.