feat: add settings for OpenTelemetry export (#2922)
## Describe your changes

Add environment settings to input OpenTelemetry endpoint and headers
![Screenshot 2024-10-30 at 13 22
26](https://github.com/user-attachments/assets/4050d7d7-f5e2-4d9d-8735-ac7a1456b371)

Note: 
- an empty endpoint is considered as disabled. I don't think adding an
extra checkmark is actually useful
- The OpenTelemetry feature is only enabled for enterprise setup or via
feature flag. I was thinking I would hide the settings if not enabled
but I think it is nice to promote the feature even if not enabled.
Editing the settings would show `OpenTelemetry export is not enabled for
this account` error. I am happy to also make a mention of it being a
paid feature in the environment setting page. Let me know
@bastienbeurier

## Checklist before requesting a review (skip if just adding/editing
APIs & templates)
- [ ] I added tests, otherwise the reason is: 
- [ ] I added observability, otherwise the reason is:
- [ ] I added analytics, otherwise the reason is: